+++ gazeplay-games/src/main/java/net/gazeplay/games/colors/AbstractGazeIndicator.java
@@ -1,4 +1,4 @@
-package net.gazeplay.components;
+package net.gazeplay.games.colors;
 
 import javafx.animation.KeyFrame;
 import javafx.animation.KeyValue;
