+++ gazeplay-games/src/main/java/net/gazeplay/games/colors/GazeFollowerIndicator.java
@@ -7,7 +7,6 @@ import javafx.scene.shape.Polygon;
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.IGameContext;
 import net.gazeplay.commons.gaze.devicemanager.GazeEvent;
-import net.gazeplay.components.AbstractGazeIndicator;
 
 @Slf4j
 public class GazeFollowerIndicator extends AbstractGazeIndicator {
