+++ gazeplay-games/src/main/java/net/gazeplay/games/colors/IGazeProgressIndicator.java
@@ -1,4 +1,4 @@
-package net.gazeplay.components;
+package net.gazeplay.games.colors;
 
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
