+++ gazeplay-games/src/main/java/net/gazeplay/games/mediaPlayer/MediaButton.java
@@ -9,7 +9,7 @@ import javafx.scene.input.MouseEvent;
 import lombok.Getter;
 import lombok.Setter;
 import net.gazeplay.commons.gaze.devicemanager.GazeEvent;
-import net.gazeplay.components.AbstractGazeIndicator;
+import net.gazeplay.games.colors.AbstractGazeIndicator;
 import net.gazeplay.components.StackPaneButton;
 
 import java.io.File;
