+++ gazeplay-games/src/test/java/net/gazeplay/games/mediaPlayer/MediaButtonTest.java
@@ -3,7 +3,7 @@ package net.gazeplay.games.mediaPlayer;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.scene.image.ImageView;
-import net.gazeplay.components.AbstractGazeIndicator;
+import net.gazeplay.games.colors.AbstractGazeIndicator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
